:recycle: split project to multi modules
