+++ common/src/main/kotlin/org/jmonster/codegenerator/common/model/dto/common/RestDto.kt
@@ -1,0 +1,3 @@
+package org.jmonster.codegenerator.common.model.dto.common
+
+interface RestDto
