+++ common/src/main/kotlin/org/jmonster/codegenerator/common/model/dto/entity/Column.kt
@@ -1,0 +1,9 @@
+package org.jmonster.codegenerator.common.model.dto.entity
+
+data class Column(
+    val key: String,
+    val name: String,
+    val type: String,
+    val nullable: String,
+    val description: String,
+)
