+++ runner/src/main/kotlin/org/jmonster/codegenerator/runner/JmonsterRunnerApplication.kt
@@ -1,0 +1,11 @@
+package org.jmonster.codegenerator.runner
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class JmonsterRunnerApplication
+
+fun main(args: Array<String>) {
+	runApplication<JmonsterRunnerApplication>(*args)
+}
