+++ settings.gradle.kts
@@ -1,1 +1,5 @@
 rootProject.name = "jmonster"
+include("common")
+include("web")
+include("runner")
+include("common")
