--- src/main/kotlin/org/jmonster/codegenerator/config/WenConfig.kt
@@ -1,18 +1,0 @@
-package org.jmonster.codegenerator.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class WenConfig : WebMvcConfigurer {
-
-    override fun addCorsMappings(registry: CorsRegistry) {
-        registry.addMapping("/**")
-            // TODO 確認環境
-            .allowedOrigins("http://localhost:4200", "http://sit_css")
-            .allowedMethods("GET", "POST", "PUT", "DELETE")
-            .allowedHeaders("*")
-            .allowCredentials(true)
-    }
-}
