--- src/main/kotlin/org/jmonster/codegenerator/controller/HomeController.kt
@@ -1,13 +1,0 @@
-package org.jmonster.codegenerator.controller
-
-import org.springframework.stereotype.Controller
-import org.springframework.web.bind.annotation.GetMapping
-
-@Controller
-class HomeController {
-
-    @GetMapping("/")
-    fun index(): String? {
-        return "index.html"
-    }
-}
