--- src/main/kotlin/org/jmonster/codegenerator/model/dto/common/RestDto.kt
@@ -1,3 +1,0 @@
-package org.jmonster.codegenerator.model.dto.common
-
-interface RestDto
