--- src/main/kotlin/org/jmonster/codegenerator/model/dto/entity/Column.kt
@@ -1,9 +1,0 @@
-package org.jmonster.codegenerator.model.dto.entity
-
-data class Column(
-    val key: String,
-    val name: String,
-    val type: String,
-    val nullable: String,
-    val description: String,
-)
