+++ web/src/main/kotlin/org/jmonster/codegenerator/web/JmonsterWebApplication.kt
@@ -1,0 +1,11 @@
+package org.jmonster.codegenerator.web
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class JmonsterApplication
+
+fun main(args: Array<String>) {
+	runApplication<JmonsterApplication>(*args)
+}
