+++ web/src/main/kotlin/org/jmonster/codegenerator/web/config/WebConfig.kt
@@ -1,0 +1,18 @@
+package org.jmonster.codegenerator.web.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.config.annotation.CorsRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+@Configuration
+class WebConfig : WebMvcConfigurer {
+
+    override fun addCorsMappings(registry: CorsRegistry) {
+        registry.addMapping("/**")
+            // TODO 確認環境
+            .allowedOrigins("http://localhost:4200", "http://sit_css")
+            .allowedMethods("GET", "POST", "PUT", "DELETE")
+            .allowedHeaders("*")
+            .allowCredentials(true)
+    }
+}
